+++ src/actions/getContract.ts
@@ -11,6 +11,7 @@ import type {
 } from 'abitype'
 
 import type { Account } from '../accounts/types.js'
+import type { Client } from '../clients/createClient.js'
 import type { Transport } from '../clients/transports/createTransport.js'
 import type { Chain } from '../types/chain.js'
 import type {
@@ -30,7 +31,6 @@ import type {
 } from '../types/utils.js'
 
 import type { ErrorType } from '../errors/utils.js'
-import type { Client } from '../index.js'
 import {
   type CreateContractEventFilterParameters,
   type CreateContractEventFilterReturnType,